refactor(meetings): promote to directory, split tests (#1266)
- src/meetings.rs: 556 LOC → src/meetings/mod.rs (239 LOC)
- New src/meetings/tests.rs (317 LOC)

Pure refactor: 22 tests pass.
